docs(slider): give RangeSlider its own documentation under derivatives/
RangeSlider previously only appeared as sections inside the Slider docs, so
it had no navigable entry of its own and was undiscoverable in the docs site.

Add a standalone, first-class doc set for RangeSlider (Overview, Guidelines,
Implementation, Accessibility tabs + consumer implementation tests) with its
own menu entry (Components > Inputs > Range Slider), and extract the
RangeSlider-specific content out of the Slider docs, replacing it with
cross-links.

Relocate everything RangeSlider-related into a self-contained
derivatives/range-slider/ subfolder (component, stories, types, index, docs),
mirroring the feat/skeleton derivatives convention. RangeSliderProps now lives
with the derivative and is built on the exported SliderSharedProps; the public
API (@commercetools/nimbus exports) is unchanged.
