+++ packages/nimbus/src/components/slider/derivatives/range-slider/index.ts
@@ -1,0 +1,2 @@
+export { RangeSlider } from "./range-slider";
+export type { RangeSliderProps } from "./range-slider.types";
