+++ packages/nimbus/src/components/slider/index.ts
@@ -1,3 +1,7 @@
 export { Slider } from "./slider";
-export { RangeSlider } from "./range-slider";
-export type { SliderProps, RangeSliderProps } from "./slider.types";
+export type { SliderProps } from "./slider.types";
+
+// Derivatives — spin-offs built on the base Slider, each self-contained under
+// `derivatives/`.
+export { RangeSlider } from "./derivatives/range-slider";
+export type { RangeSliderProps } from "./derivatives/range-slider";
