+++ packages/nimbus/src/components/slider/slider.guidelines.mdx
@@ -28,25 +28,24 @@ accessible name, even though it draws no visible label of its own.
   repeats it. A `FormField.Description` summarizing the selection elsewhere
   in the form (e.g. a range's bounds) is a separate, permanent piece of
   context and is fine to keep.
-- **Use `RangeSlider` for bounded ranges.** Its two handles can't cross, so it
-  guarantees the lower value never exceeds the upper one — useful for filters
-  like price or distance ranges.
+- **Reach for the [Range Slider](/components/inputs/range-slider) for bounded
+  ranges.** Its two handles can't cross, so it guarantees the lower value never
+  exceeds the upper one — useful for filters like price or distance ranges.
 - **Disable, don't hide.** When a slider's value is temporarily not
   applicable, disable it with `isDisabled` rather than removing it from the
   layout, so surrounding controls don't shift.
 
 ### Usage
 
-A slider lets users choose a value — or, with `RangeSlider`, a range — by
-dragging a handle along a track, best for values that are set by feel rather
-than typed precisely.
+A slider lets users choose a value by dragging a handle along a track, best for
+values that are set by feel rather than typed precisely. To choose a range
+instead, use the [Range Slider](/components/inputs/range-slider).
 
 > [!TIP]\
 > When to use
 >
 > - Selecting a value from a continuous or stepped numeric scale by feel, such
 >   as volume, zoom, opacity, or brightness.
-> - Filtering by a numeric range, such as price, distance, or rating.
 > - The exact number matters less than its position relative to the minimum
 >   and maximum.
 
@@ -73,8 +72,8 @@ unexpected jumps or hidden values erode trust in the control.
 >   `aria-label`.
 > - Set `minValue`, `maxValue`, and a `step` that make sense for the data.
 > - Use `showTicks` when the discrete stops themselves are meaningful.
-> - Use `RangeSlider` when both a lower and upper bound need to be set
->   together.
+> - Reach for the [Range Slider](/components/inputs/range-slider) when both a
+>   lower and upper bound need to be set together.
 > - Let the disabled state show when the value can't currently be changed.
 
 ```jsx live
